+++ tests/testthat/test-add_reference_rows.R
@@ -115,28 +115,32 @@ test_that("test tidy_add_reference_rows() checks", {
 
   # warning if applied twice
   expect_message(
-    mod %>% tidy_and_attach() %>%
+    mod %>%
+      tidy_and_attach() %>%
       tidy_add_reference_rows() %>%
       tidy_add_reference_rows()
   )
 
   # message if applied after tidy_add_term_labels()
   expect_message(
-    mod %>% tidy_and_attach() %>%
+    mod %>%
+      tidy_and_attach() %>%
       tidy_add_term_labels() %>%
       tidy_add_reference_rows()
   )
 
   # message if applied after tidy_add_n()
   expect_message(
-    mod %>% tidy_and_attach() %>%
+    mod %>%
+      tidy_and_attach() %>%
       tidy_add_n() %>%
       tidy_add_reference_rows()
   )
 
   # error if applied after tidy_add_header_rows()
   expect_error(
-    mod %>% tidy_and_attach() %>%
+    mod %>%
+      tidy_and_attach() %>%
       tidy_add_header_rows() %>%
       tidy_add_reference_rows()
   )
@@ -236,7 +240,7 @@ test_that("tidy_add_reference_rows() works with lme4::glmer", {
     tidy_add_reference_rows()
 
   expect_equal(
-    res[res$reference_row & !is.na(res$reference_row),]$effect,
+    res[res$reference_row & !is.na(res$reference_row), ]$effect,
     "fixed"
   )
 })